requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-crispy-forms==1.13.0
django-froala-editor==4.0.9
django-import-export==2.7.1
django-js-asset==1.2.2
django-tinymce==3.4.0
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
mysql-connector-python==8.0.28
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
protobuf==3.20.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
tablib==3.2.0
typing-extensions==4.0.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
